BACKPORT: drm/file: Add fdinfo helper for printing regions with prefix
This is motivated by the desire of some drivers (eg. Panthor) to print the
size of internal memory regions with a prefix that reflects the driver
name, as suggested in the previous documentation commit.

That means adding a new argument to print_size and making it available for
DRM users.

Cc: Tvrtko Ursulin <tursulin@ursulin.net>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@igalia.com>
Signed-off-by: Adrián Larumbe <adrian.larumbe@collabora.com>
Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20250130172851.941597-3-adrian.larumbe@collabora.com
(cherry picked from commit af6c2b7c46e16701fba44a21326cb634786e3e71)
[olv: fix minor conflicts due to missing fd265d9e0c335 ("drm: add
 drm_memory_stats_is_zero") and bebf2ebd70f21 ("drm: make drm-active-
 stats optional")]

Bug: 439665327

Change-Id: I5f4b30503fac6ff494737d4a69ea4185b0d9421c
Signed-off-by: Chia-I Wu <olv@google.com>