ANDROID: iommu/arm-smmu-v3-kvm: Add support for reporting IOMMU faults to clients
IOMMU clients can register callbacks to be invoked when an IOMMU fault
occurs. The IOMMU driver that detects the fault can report faults to
an IOMMU client driver via the report_iommu_fault() function. Therefore,
add support in the SMMU driver to do so.

This is based on the upstream implementation of event handling in the
SMMUv3 driver in Linux kernel version 6.15-rc5, except it is trimmed
down to only fetch the fields necessary for the report_iommu_fault()
function.

Bug: 369229886
Bug: 437825935
Change-Id: I9ac0f8b478305450ccff5be8b56981d2deb6ebe4
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>
Signed-off-by: Daniel Mentz <danielmentz@google.com>