ANDROID: zram: Add ioctl infrastructure
Add the basic infrastructure for handling ioctl commands in the zram
driver. This provides a channel for device-specific operations between
userspace and the driver. Introduce ZRAM_ANDROID_IOCTL config.

This change is a preparatory step for introducing more advanced zram
features. The new ioctl interface will be used to implement per-process
page operations in a subsequent patch.

Bug: 421065577
Change-Id: I9d34b88fb4d98df0bf7e5b04d4820bb6ef0a0c24
Signed-off-by: Richard Chang <richardycc@google.com>