ANDROID: Re-add cfs_rq.h_nr_running element for KMI compatability
In commit a30857155536 ("UPSTREAM: sched/fair: Rename
h_nr_running into h_nr_queued"), we pulled in a change that
renamed a structure element as part of the KMI thaw. This
shouldn't have been done as the KMI thaw requires no changes
that could break existing source code.

However, there are a number of changes in -stable that expect
this rename so it would be nice to preserve it.

So use ANDROID_KABI_REPLACE to preserve the h_nr_running name
along with the h_nr_queued.

With this we do risk some confusion as part of the rename a new
element h_nr_runnable was added in commit f223aa8ea2902
("UPSTREAM: sched/fair: Add new cfs_rq.h_nr_runnable") as part
of the kmi thaw.

But for existing code the semantics of h_nr_running should match
h_nr_queued.

Bug: 446253608
Change-Id: I9019396af6af9bc3af4cb1a828a2d32bf992d51c
Signed-off-by: John Stultz <jstultz@google.com>