ANDROID: mm: Export get_swap_device for zram module
The zram module requires access to swap_info_struct to validate
swap entries for the ioctl writeback operations.

This validation ensures that the swap entries belong to the correct
zram device, and uses the get/put_swap_device() functions to prevent the
swapoff/swapon race condition. Export the symbol to make it available to
access by the zram module.

Bug: 447017656
Change-Id: Ida25745866f94f43c261c5da76e6b681ae35f6ce
Signed-off-by: Richard Chang <richardycc@google.com>