Revert "ANDROID: vendor_hooks:add restricted hook for logbuf"
This reverts commit 6fa673a17109fd8d029af9216240eaead3ccdd6e.

The hook android_rvh_logbuf is not used by any vendor, so remove it to
help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 364443023
Cc: Xandy.Xiong <xiongliang@xiaomi.com>
Change-Id: I2faa2c99274e5b77c03c73d402ad0482e9845953
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>