Revert "ANDROID: sched: Add vendor hooks for find_energy_efficient_cpu"
This reverts commit b8ea476fb4cb33062eded56d93ecaa49e32040cb.

The hook android_rvh_find_energy_efficient_cpu is not used by any
vendor, so remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 170507310
Bug: 265387484
Bug: 329341111
Bug: 377187885
Cc: Rick Yiu <rickyiu@google.com>
Cc: Will McVicker <willmcvicker@google.com>
Cc: Han Lin <han.lin@mediatek.com>
Change-Id: Ia904610132a6974fb461152ec3ad2c7a20e7a215
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>