Revert "ANDROID: Re-apply vendor hooks for information of blocked tasks"
This reverts a part of commit fb27e5d8ba0e81e2c6d618465b4c0cbdbac6c46c.

The hook android_vh_rtmutex_wait_finish is not used by any vendor, so
remove it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 271799327
Cc: Sangmoon Kim <sangmoon.kim@samsung.com>
Change-Id: I5a9a38e635ad166416722787bb2b605bf3ce005a
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>