ANDROID: zram: Stop using kernel_read_file_from_path
The kernel_read_file_from_path() function is on GKI's symbol blocklist,
which means it's currently impossible to build zram as an in-tree
module. Update zram_drv.c to use read_comp_algo_dictionary(), which is
a wrapper function around kernel_read_file_from_path() that was added to
unblock vendors' out-of-tree zram modules.

Bug: 443091113
Change-Id: I6b44d7527d207d81ff986a23c6a1e3d808b485f2
Signed-off-by: David Stevens <stevensd@google.com>