Revert "ANDROID: dma-buf: heaps: Allow cma heaps to be configured as a module"
This reverts commit 70735d9b05a1e9c755c17e26785a0de21e01c6bf.

The export of dma_heap_get_drvdata() is no longer needed as upstream
does it in commit as a6ccb03fb773 ("dma-buf: dma-heap: export declared
functions"). Also, there is no trace of us setting DMABUF_HEAPS_CMA=m
ever, so drop the "tristate" support for it. This can be brought back if
there is really a need for it.

Bug: 155218010
Cc: John Stultz <jstultz@google.com>
Change-Id: I8ea136655d6b61219fed36786c2f73fe7a624ca9
Signed-off-by: Carlos Llamas <cmllamas@google.com>