ANDROID: KVM: arm64: sync_icache_alias for pKVM module symbols reloc
Calls to symbols exported from other pKVM modules emit a relocation
that needs to be patched. Just like with other relocations the code is
being patched and icache maintenance must be performed.

Luckily, sync_icache_alias is already called in
kvm_apply_hyp_module_relocations() so we can reuse it.

Bug: 455467611
Bug: 333671007
Bug: 357781595
Change-Id: Ia7260c714d56cb7a411850eddefd3d4bf19734b4
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>